+++ core/src/server/web_impl/controller/WebController.hpp
@@ -257,10 +257,15 @@ class WebController : public oatpp::web::server::api::ApiController {
 
     ADD_CORS(ShowTables)
 
-    ENDPOINT("GET", "/tables", ShowTables, QUERY(Int64, offset, "offset"), QUERY(Int64, page_size, "page_size")) {
+//    ENDPOINT("GET", "/tables", ShowTables, QUERY(Int64, offset, "offset"), QUERY(Int64, page_size, "page_size")) {
+    ENDPOINT("GET", "/tables", ShowTables, REQUEST(
+        const std::shared_ptr<IncomingRequest>&, request)) {
         WebRequestHandler handler = WebRequestHandler();
         handler.RegisterRequestHandler(::milvus::server::RequestHandler());
         auto response_dto = TableListFieldsDto::createShared();
+        auto offset = request->getQueryParameter("offset", "0");
+        auto page_size = request->getQueryParameter("page_size", "10");
+
         auto status_dto = handler.ShowTables(offset, page_size, response_dto);
         std::shared_ptr<OutgoingResponse> response;
         if (0 == status_dto->code->getValue()) {
@@ -309,8 +314,6 @@ class WebController : public oatpp::web::server::api::ApiController {
     ENDPOINT_INFO(DropTable) {
         info->summary = "Drop table";
 
-        info->pathParams.add<String>("table_name");
-
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_204, "application/json");
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
@@ -341,11 +344,11 @@ class WebController : public oatpp::web::server::api::ApiController {
     ENDPOINT_INFO(CreateIndex) {
         info->summary = "Create index";
 
-        info->pathParams.add<String>("table_name");
-
         info->addConsumes<IndexRequestDto::ObjectWrapper>("application/json");
+
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_201, "application/json");
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
+        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
     }
 
     ADD_CORS(CreateIndex)
@@ -355,9 +358,12 @@ class WebController : public oatpp::web::server::api::ApiController {
         auto handler = WebRequestHandler();
         handler.RegisterRequestHandler(::milvus::server::RequestHandler());
         auto status_dto = handler.CreateIndex(table_name, body);
+        auto code = status_dto->code->getValue();
 
-        if (0 == status_dto->code->getValue()) {
+        if (0 == code) {
             return createDtoResponse(Status::CODE_201, status_dto);
+        } else if (StatusCode::TABLE_NOT_EXISTS == code) {
+            return createDtoResponse(Status::CODE_404, status_dto);
         } else {
             return createDtoResponse(Status::CODE_400, status_dto);
         }
@@ -431,6 +437,7 @@ class WebController : public oatpp::web::server::api::ApiController {
 
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_201, "application/json");
         info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
+        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
     }
 
     ADD_CORS(CreatePartition)
@@ -440,9 +447,12 @@ class WebController : public oatpp::web::server::api::ApiController {
         auto handler = WebRequestHandler();
         handler.RegisterRequestHandler(::milvus::server::RequestHandler());
         auto status_dto = handler.CreatePartition(table_name, body);
+        auto code = status_dto->code->getValue();
 
-        if (0 == status_dto->code->getValue()) {
+        if (0 == code) {
             return createDtoResponse(Status::CODE_201, status_dto);
+        } else if (StatusCode::TABLE_NOT_EXISTS == code) {
+            return createDtoResponse(Status::CODE_404, status_dto);
         } else {
             return createDtoResponse(Status::CODE_400, status_dto);
         }
@@ -555,37 +565,6 @@ class WebController : public oatpp::web::server::api::ApiController {
         }
     }
 
-    ENDPOINT_INFO(InsertBin) {
-        info->summary = "Insert vectors";
-
-        info->pathParams.add<String>("table_name");
-
-        info->addConsumes<InsertRequestDto::ObjectWrapper>("application/json");
-
-        info->addResponse<VectorIdsDto::ObjectWrapper>(Status::CODE_201, "application/json");
-        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
-        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
-    }
-
-    ADD_CORS(InsertBin)
-
-    ENDPOINT("POST", "/tables/{table_name}/vectors/bin", InsertBin,
-             PATH(String, table_name), BODY_DTO(InsertRequestDto::ObjectWrapper, body)) {
-        auto ids_dto = VectorIdsDto::createShared();
-        WebRequestHandler handler = WebRequestHandler();
-        handler.RegisterRequestHandler(::milvus::server::RequestHandler());
-        auto status_dto = handler.Insert(table_name, body, ids_dto);
-
-        int64_t code = status_dto->code->getValue();
-        if (0 == code) {
-            return createDtoResponse(Status::CODE_201, ids_dto);
-        } else if (StatusCode::TABLE_NOT_EXISTS == code) {
-            return createDtoResponse(Status::CODE_404, status_dto);
-        } else {
-            return createDtoResponse(Status::CODE_400, status_dto);
-        }
-    }
-
     ENDPOINT_INFO(Search) {
         info->summary = "Search";
 
@@ -616,35 +595,72 @@ class WebController : public oatpp::web::server::api::ApiController {
         }
     }
 
-    ENDPOINT_INFO(SearchBin) {
-        info->summary = "Search for binary vectors";
-
-        info->pathParams.add<String>("table_name");
-
-        info->addConsumes<SearchBinRequestDto::ObjectWrapper>("application/json");
-
-        info->addResponse<TopkResultsDto::ObjectWrapper>(Status::CODE_200, "application/json");
-        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
-        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
-    }
-
-    ADD_CORS(SearchBin)
-
-    ENDPOINT("PUT", "/tables/{table_name}/vectors/bin", SearchBin,
-             PATH(String, table_name), BODY_DTO(SearchRequestDto::ObjectWrapper, body)) {
-        auto results_dto = TopkResultsDto::createShared();
-        WebRequestHandler handler = WebRequestHandler();
-        handler.RegisterRequestHandler(::milvus::server::RequestHandler());
-        auto status_dto = handler.Search(table_name, body, results_dto);
-        int64_t code = status_dto->code->getValue();
-        if (0 == code) {
-            return createDtoResponse(Status::CODE_200, results_dto);
-        } else if (StatusCode::TABLE_NOT_EXISTS == code) {
-            return createDtoResponse(Status::CODE_404, status_dto);
-        } else {
-            return createDtoResponse(Status::CODE_400, status_dto);
-        }
-    }
+//    ADD_CORS(VectorsBinOptions)
+//
+//    ENDPOINT("OPTIONS", "/tables/{table_name}/vectors/bin", VectorsBinOptions) {
+//        return createResponse(Status::CODE_204, "No Content");
+//    }
+//
+//    ENDPOINT_INFO(InsertBin) {
+//        info->summary = "Insert vectors";
+//
+//        info->pathParams.add<String>("table_name");
+//
+//        info->addConsumes<InsertBinRequestDto::ObjectWrapper>("application/json");
+//
+//        info->addResponse<VectorIdsDto::ObjectWrapper>(Status::CODE_201, "application/json");
+//        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
+//        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
+//    }
+//
+//    ADD_CORS(InsertBin)
+//
+//    ENDPOINT("POST", "/tables/{table_name}/vectors/bin", InsertBin,
+//             PATH(String, table_name), BODY_DTO(InsertBinRequestDto::ObjectWrapper, body)) {
+//        auto ids_dto = VectorIdsDto::createShared();
+//        WebRequestHandler handler = WebRequestHandler();
+//        handler.RegisterRequestHandler(::milvus::server::RequestHandler());
+//        auto status_dto = handler.Insert<InsertBinRequestDto::ObjectWrapper>(table_name, body, ids_dto);
+//
+//        int64_t code = status_dto->code->getValue();
+//        if (0 == code) {
+//            return createDtoResponse(Status::CODE_201, ids_dto);
+//        } else if (StatusCode::TABLE_NOT_EXISTS == code) {
+//            return createDtoResponse(Status::CODE_404, status_dto);
+//        } else {
+//            return createDtoResponse(Status::CODE_400, status_dto);
+//        }
+//    }
+//
+//    ENDPOINT_INFO(SearchBin) {
+//        info->summary = "Search for binary vectors";
+//
+//        info->pathParams.add<String>("table_name");
+//
+//        info->addConsumes<SearchBinRequestDto::ObjectWrapper>("application/json");
+//
+//        info->addResponse<TopkResultsDto::ObjectWrapper>(Status::CODE_200, "application/json");
+//        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_400, "application/json");
+//        info->addResponse<StatusDto::ObjectWrapper>(Status::CODE_404, "application/json");
+//    }
+//
+//    ADD_CORS(SearchBin)
+//
+//    ENDPOINT("PUT", "/tables/{table_name}/vectors/bin", SearchBin,
+//             PATH(String, table_name), BODY_DTO(SearchBinRequestDto::ObjectWrapper, body)) {
+//        auto results_dto = TopkResultsDto::createShared();
+//        WebRequestHandler handler = WebRequestHandler();
+//        handler.RegisterRequestHandler(::milvus::server::RequestHandler());
+//        auto status_dto = handler.Search<SearchBinRequestDto::ObjectWrapper>(table_name, body, results_dto);
+//        int64_t code = status_dto->code->getValue();
+//        if (0 == code) {
+//            return createDtoResponse(Status::CODE_200, results_dto);
+//        } else if (StatusCode::TABLE_NOT_EXISTS == code) {
+//            return createDtoResponse(Status::CODE_404, status_dto);
+//        } else {
+//            return createDtoResponse(Status::CODE_400, status_dto);
+//        }
+//    }
 
     ENDPOINT_INFO(SystemMsg) {
         info->summary = "Command";